Move utilities to valid package
