+++ ThirdPartyAdapters/zucks/zucks/src/main/java/com/google/ads/mediation/zucks/AdMobUtil.java
@@ -1,4 +1,4 @@
-package net.zucks.admob;
+package com.google.ads.mediation.zucks;
 
 import android.os.Bundle;
