+++ ThirdPartyAdapters/zucks/zucks/src/main/java/com/google/ads/mediation/zucks/BaseMediationAdapter.java
@@ -1,4 +1,4 @@
-package net.zucks.admob;
+package com.google.ads.mediation.zucks;
 
 import android.content.Context;
