+++ ThirdPartyAdapters/zucks/zucks/src/test/java/com/google/ads/mediation/zucks/UniversalInterstitialListenerTest.kt
@@ -1,4 +1,4 @@
-package net.zucks.admob
+package com.google.ads.mediation.zucks
 
 import io.mockk.mockk
 import io.mockk.verify
